refactor(types): extract context handoff DTOs
